update imports and rename folder with icons
